Add all of the Compcon font symbols to the document.  Some a renamed for legibility and consistency with Compcon's naming scheme. (#6)
This may break some documents, though I tried to keep the ones named in the document the same.  Some old keys are available for backward compatibility.
The one real change is the I had the wrong symbols for NPC Tiers, those were the symbols for the Talent Tiers, so they have been corrected, and the Talent Tiers are now called \CCRankOne, ect.